refactor: addressing stricter IDE quality checks [incremental]

+++ src/main/java/com/gotreaux/aoc/puzzles/year2016/day3/Triangle.java
@@ -1,8 +1,8 @@
 package com.gotreaux.aoc.puzzles.year2016.day3;
 
-record Triangle(long a, long b, long c) {
+record Triangle(int a, int b, int c) {
     Triangle {
-        if (a <= 0L || b <= 0L || c <= 0L) {
+        if (a <= 0 || b <= 0 || c <= 0) {
             throw new IllegalArgumentException(
                     "Every side of the triangle must be a positive number!");
         }
