+++ src/test/java/com/gotreaux/aoc/puzzles/year2016/day3/TriangleTest.java
@@ -15,26 +15,26 @@ class TriangleTest {
     void throwsIfNegativeSideLength() {
         RandomGenerator generator = RandomGenerator.getDefault();
 
-        long negativeArgumentIndex = generator.nextLong(1, 4);
-        long a =
-                negativeArgumentIndex == 1L
-                        ? -Math.abs(generator.nextLong())
-                        : Math.abs(generator.nextLong());
-        long b =
-                negativeArgumentIndex == 2L
-                        ? -Math.abs(generator.nextLong())
-                        : Math.abs(generator.nextLong());
-        long c =
-                negativeArgumentIndex == 3L
-                        ? -Math.abs(generator.nextLong())
-                        : Math.abs(generator.nextLong());
+        int negativeArgumentIndex = generator.nextInt(1, 4);
+        int a =
+                negativeArgumentIndex == 1
+                        ? -Math.abs(generator.nextInt())
+                        : Math.abs(generator.nextInt());
+        int b =
+                negativeArgumentIndex == 2
+                        ? -Math.abs(generator.nextInt())
+                        : Math.abs(generator.nextInt());
+        int c =
+                negativeArgumentIndex == 3
+                        ? -Math.abs(generator.nextInt())
+                        : Math.abs(generator.nextInt());
 
         assertThrows(IllegalArgumentException.class, () -> new Triangle(a, b, c));
     }
 
     @ParameterizedTest
     @MethodSource("provideValidTriangle")
-    void validTriangle(long a, long b, long c, boolean isValid) {
+    void validTriangle(int a, int b, int c, boolean isValid) {
         Triangle triangle = new Triangle(a, b, c);
 
         assertEquals(isValid, triangle.isValid());
@@ -42,12 +42,12 @@ class TriangleTest {
 
     private static Stream<Arguments> provideValidTriangle() {
         return Stream.of(
-                Arguments.of(5L, 10L, 25L, false),
-                Arguments.of(101L, 102L, 103L, true),
-                Arguments.of(201L, 202L, 203L, true),
-                Arguments.of(301L, 302L, 303L, true),
-                Arguments.of(401L, 402L, 403L, true),
-                Arguments.of(501L, 502L, 503L, true),
-                Arguments.of(601L, 602L, 603L, true));
+                Arguments.of(5, 10, 25, false),
+                Arguments.of(101, 102, 103, true),
+                Arguments.of(201, 202, 203, true),
+                Arguments.of(301, 302, 303, true),
+                Arguments.of(401, 402, 403, true),
+                Arguments.of(501, 502, 503, true),
+                Arguments.of(601, 602, 603, true));
     }
 }
